pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "owli-detector-train"
version = "0.1.0"
description = "COCO dataset tooling + Keras training + TFLite export for Android offline inference."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Owli" }]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "PyYAML>=6.0",
  "numpy>=1.24",
  "Pillow>=10.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
]
eval = [
  "pycocotools>=2.0.7",
]
keras = [
  "tensorflow>=2.15",
  "keras-cv>=0.9.0",
]

[project.scripts]
owli-train = "owli_train.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/owli_train"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
  "TMP",
  "tmp",
  ".tmp",
  "**/TMP/**",
  "**/tmp/**",
  "work",
  "outputs",
  "data",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
norecursedirs = ["tmp", "TMP", ".tmp", "work", "outputs", "data"]
